gpu-server/requirements.txt
# PyTorch（根据 CUDA 版本选择，此处为 CUDA 11.8）
# 安装时运行：pip install torch==2.1.0 torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cu118
torch==2.1.0
torchvision==0.16.0

# Transformers（模型加载）
transformers==4.40.0
accelerate==0.30.0

# 图像处理
Pillow==10.3.0
rasterio==1.3.10
numpy==1.26.4

# HTTP 客户端
requests==2.32.3

# 配置管理
python-dotenv==1.0.1

# 日志
loguru==0.7.3

# 进度条
tqdm==4.66.4
